- 将web hook脚本放到了新的目录 - 修改hook的加载方式为注册模式
